requirements.txt
pytorch==1.12.1
torchvision==0.13.1
timm==0.6.7
pillow==9.2.0
matplotlib==3.5.3
torchprofile==0.0.4
six==1.16.0
scipy==1.7.3
pandas==1.4.4